+++ changelog.md
@@ -8,6 +8,7 @@ improvements may be omitted.)_
 
 __[maintenance]__
 
+* Update certificate information for HeidiSQL installer.
 * Update certificate information for SeaMonkey installers, because installers
   are no longer signed.